Centralise sort mapping utils
